+++ src/mjlab/scripts/csv_to_npz.py
@@ -325,10 +325,11 @@ def run_sim(
         print(f"[INFO]: Motion saved to wandb registry: {REGISTRY}/{COLLECTION}")
 
         if render:
-          import mediapy as media
+          from moviepy import ImageSequenceClip
 
           print("Creating video...")
-          media.write_video("./motion.mp4", frames, fps=output_fps)
+          clip = ImageSequenceClip(frames, fps=output_fps)
+          clip.write_videofile("./motion.mp4")
 
           print("Logging video to wandb...")
           wandb.log({"motion_video": wandb.Video("./motion.mp4", format="mp4")})